+++ src/fiuba/algo3/tp2/DiamanteMaterial.java
@@ -1,5 +1,7 @@
 package fiuba.algo3.tp2;
 
+import fiuba.algo3.tp2.herramientas.PicoFino;
+
 public class DiamanteMaterial extends Material {
     public DiamanteMaterial() {
         this.durabilidad = 100;
