+++ src/fiuba/algo3/tp2/MetalMaterial.java
@@ -1,5 +1,7 @@
 package fiuba.algo3.tp2;
 
+import fiuba.algo3.tp2.herramientas.PicoPiedra;
+
 public class MetalMaterial extends Material {
     public MetalMaterial() {
         this.durabilidad = 50;
