--- src/fiuba/algo3/tp2/Pico.java
@@ -1,6 +1,0 @@
-package fiuba.algo3.tp2;
-
-import fiuba.algo3.tp2.Herramienta;
-
-public class Pico extends Herramienta {
-}
