+++ src/fiuba/algo3/tp2/PiedraMaterial.java
@@ -1,5 +1,7 @@
 package fiuba.algo3.tp2;
 
+import fiuba.algo3.tp2.herramientas.Pico;
+
 public class PiedraMaterial extends Material {
     public PiedraMaterial() {
         this.durabilidad = 30;
