+++ src/fiuba/algo3/tp2/herramientas/Herramienta.java
@@ -1,5 +1,6 @@
-package fiuba.algo3.tp2;
+package fiuba.algo3.tp2.herramientas;
 
+import fiuba.algo3.tp2.Material;
 import fiuba.algo3.tp2.estrategiasDesgaste.EstrategiaDesgaste;
 
 public class Herramienta {
