+++ src/fiuba/algo3/tp2/herramientas/Pico.java
@@ -1,0 +1,6 @@
+package fiuba.algo3.tp2.herramientas;
+
+import fiuba.algo3.tp2.herramientas.Herramienta;
+
+public class Pico extends Herramienta {
+}
