+++ src/fiuba/algo3/tp2/herramientas/PicoMetal.java
@@ -1,4 +1,4 @@
-package fiuba.algo3.tp2;
+package fiuba.algo3.tp2.herramientas;
 
 import fiuba.algo3.tp2.estrategiasDesgaste.EstrategiaDesgasteAbrupta;
