+++ test/fiuba/algo3/tp2/HachaTest.java
@@ -1,6 +1,9 @@
 package fiuba.algo3.tp2;
 import static org.junit.Assert.*;
 
+import fiuba.algo3.tp2.herramientas.HachaDeMadera;
+import fiuba.algo3.tp2.herramientas.HachaDeMetal;
+import fiuba.algo3.tp2.herramientas.HachaDePiedra;
 import org.junit.Test;
 
 public class HachaTest {
