+++ test/fiuba/algo3/tp2/PicoMaderaTest.java
@@ -1,5 +1,6 @@
 package fiuba.algo3.tp2;
 
+import fiuba.algo3.tp2.herramientas.PicoMadera;
 import org.junit.Assert;
 import org.junit.Test;
